split CLIOptions interface to own file
